Now if consolidar.sh exists, the field is copied in HOME directory and executed in background, using consolidar.log to visualize errors

+++ crear_entorno.sh
@@ -5,8 +5,19 @@ EPNro1_DIR="$HOME/EPNro1"
 echo "Creando entorno en $EPNro1_DIR..."
 
 if mkdir -p "$EPNro1_DIR"/{entrada,salida,procesado}; then
-    echo "Copiando consolidar.sh en HOME y ejecuandolo en background..."
-    cp consolidar.sh "$EPNro1_DIR/"
+    touch "$EPNro1_DIR/salida/FILENAME.txt "
+
+    if [ -f "consolidar.sh" ]; then
+        echo "Copiando consolidar.sh en HOME/EPNro1 y ejecutandolo en background..."
+        cp consolidar.sh "$EPNro1_DIR/"
+
+        chmod +x "$EPNro1_DIR/consolidar.sh"
+
+        "$EPNro1_DIR/consolidar.sh" > "$EPNro1_DIR/consolidar.log" 2>&1 &
+    else
+        echo "No se encontró consolidar.sh"
+    fi
+
 else
     echo "Error al crear el entorno"
 fi
